Non-linked variants of acdh:hasCoverage and acdh:hasContributor added
Non-linked variants were created only for optional and recommended
properties as for required ones we would be unable to properly express
the cardinality requirement. We can also look at it in a way that if a
property is required, it's so imporant we want it to be LOD and don't
allow non-linked data.